moving class definitions to own files, fix CPY_VERSION references in modules
